+++ Print/PrintAtTheEnd.cpp
@@ -6,14 +6,14 @@
 
 using namespace chrono;
 
-void PrintAtTheEnd::logPrime(int prime, int threadId) {
+void PrintAtTheEnd::logPrime(unsigned long int prime, unsigned int threadId) {
     lock_guard<mutex> lock(printMutex);
     allPrimes.push_back(prime);
 
     this_thread::sleep_for(milliseconds(100));  // Delay to simulate processing time
 }
 
-void PrintAtTheEnd::printPrimes(int prime, int threadId)
+void PrintAtTheEnd::printPrimes(unsigned long int prime, unsigned int threadId)
 {
     lock_guard<mutex> lock(printMutex);
     for (auto prime : allPrimes) {
